move fitting callbacks into fitting.__init__
