+++ src/ibex_bluesky_core/callbacks/__init__.py
@@ -20,10 +20,7 @@ from ibex_bluesky_core.callbacks._utils import get_default_output_path
 from ibex_bluesky_core.callbacks.file_logger import (
     HumanReadableFileCallback,
 )
-from ibex_bluesky_core.callbacks.fitting import FitMethod, LiveFit
-from ibex_bluesky_core.callbacks.fitting.livefit_logger import (
-    LiveFitLogger,
-)
+from ibex_bluesky_core.callbacks.fitting import FitMethod, LiveFit, LiveFitLogger
 from ibex_bluesky_core.callbacks.plotting import LivePlot
 
 logger = logging.getLogger(__name__)
